two problem should solve
1: the value that get_rdtsc() return is long long type,but
   cdll.LoadLibrary is return int, so we can not get the
   correct value

2: function of sendto is needed bytes,so i need to chage the
   correct format

+++ rdtsc.c
@@ -1,10 +1,11 @@
-typedef unsigned long long __u64;
-typedef unsigned long __u32;
-int get_rdtsc(){
+typedef long long __u64;
+typedef long __u32;
+__u64 get_rdtsc(){
     __u32 hi,lo;
     __asm__ __volatile__ 
     (
         "rdtsc" : "=a" (lo),"=d"(hi)
     );
+
     return (__u64) hi<<32|lo;
 }
